use timer in iirfilter
results for me: best out of twenty for mouse.png
0.0721834 for iirfilter
0.0259404 for test_seq_deriche_2d
But I think iirfilter does calculate something for the boundary which the
other version does not do - on the other hand test_seq_deriche does
horizontal _and_ vertical pass which I think iirfilter does not? Difference
is probably due to the fact that test_seq_deriche adapts its computation to
the order of the deriche approximation.